@@ -2,11 +2,8 @@
   "name": "clc.wtf",
   "version": "0.0.1",
   "description": "My first TS project",
-  "main": "index.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/dniwe-team/clc.wtf.git"
-  },
+  "main": "src/index.js",
+  "repository": "git+https://github.com/dniwe-team/clc.wtf.git",
   "author": "TheDeGor <degor1983@gmail.com>",
   "license": "MIT",
   "bugs": {
@@ -14,25 +11,34 @@
   },
   "homepage": "https://github.com/dniwe-team/clc.wtf#readme",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "yarn eslint 'src/**/*.js'",
+    "test": "ts-node -T src/index.ts",
+    "lint": "yarn eslint 'src/**/*.ts'",
     "dbup": "db-migrate up",
-    "dbdown": "db-migrate down"
+    "dbdown": "db-migrate down",
+    "dev": "ts-node-dev --respawn --transpile-only src/index.ts",
+    "tsc": "tsc",
+    "tsc:w": "tsc -w"
   },
   "dependencies": {
-    "body-parser": "^1.19.0",
     "config": "^3.3.1",
     "db-migrate-pg": "^1.2.2",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "tsc": "^1.20150623.0",
-    "typescript": "^3.8.3"
+    "tsc": "^1.20150623.0"
   },
   "devDependencies": {
+    "@types/config": "0.0.36",
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.7",
+    "@types/node": "^14.0.27",
     "@typescript-eslint/eslint-plugin": "^2.31.0",
     "@typescript-eslint/parser": "^2.31.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
-    "eslint-plugin-import": "^2.20.2"
+    "eslint-plugin-import": "^2.22.0",
+    "ts-node": "^8.10.2",
+    "ts-node-dev": "^1.0.0-pre.56",
+    "tsconfig": "^7.0.0",
+    "typescript": "^3.9.7"
   }
 }
